requirements.txt
streamlit
openpyxl
pillow
qrcode
